fix: repair syntax errors, wire all modules, enforce 400-line limit
- Register all 8 utils modules + project_init_detect in lib.rs
- Split project_init.rs (468->305 lines) by extracting detect helpers
- Extract scenario.rs tests to src/tests/ (513->369 lines)
