+++ crates/amplihack-utils/src/lib.rs
@@ -10,12 +10,17 @@
 //! - [`project_init`] — Project initialization and `PROJECT.md` management
 //! - [`claude_md`] — `CLAUDE.md` preservation and version management
 //! - [`cleanup`] — Cleanup registry and handler for tracked temporary paths
+//! - [`claude_cli`] — Claude CLI binary detection, installation, version checking
+//! - [`prerequisites`] — Tool detection and installation guidance
 
+pub mod claude_cli;
 pub mod claude_md;
 pub mod cleanup;
 pub mod defensive;
+pub mod prerequisites;
 pub mod process;
 pub mod project_init;
+pub(crate) mod project_init_detect;
 pub mod slugify;
 
 // Re-export the most commonly used items at crate root.
